lm_linear

lm_linear

mcmc_linear(
  y,
  X,
  plot_idx,
  tree_idx,
  params,
  config = NULL,
  verbose = FALSE,
  n_chain = 1
)

Arguments

y
A vector of length n of tree ring increments on a log scale

X
A n times p matrix of covariate values

plot_idx
A vector of length n that indicates which plot each observation was sampled from

tree_idx
A vector of length n indicating which tree the observation came from

params
is a list of parameter settings. The list params must contain the following values:

n_adapt: A positive integer number of adaptive MCMC iterations.

n_mcmc: A positive integer number of total MCMC iterations post adaptation.

n_thin: A positive integer number of MCMC iterations per saved sample.

n_message: A positive integer number of frequency of iterations to output a progress message. For example, n_message = 50 outputs progress messages every 50 iterations.

config
is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.

verbose
Should verbose output be printed? Typically this is only useful for troubleshooting.

n_chain
is the MCMC chain id. The default is 1.

Value

A MCMC object that is the result of fitting the linear regression model